+++ packages/contracts/contracts/Interfaces/IRecoveryModeGracePeriod.sol
@@ -3,7 +3,7 @@ pragma solidity 0.8.17;
 
 // Interface for State Updates that can trigger RM Liquidations
 interface IRecoveryModeGracePeriod {
-    event TCRNotified(uint TCR); /// NOTE: Mostly for debugging to ensure synch
+    event TCRNotified(uint256 TCR); /// NOTE: Mostly for debugging to ensure synch
 
     // NOTE: Ts is implicit in events (it's added by GETH)
     event GracePeriodStart();
