+++ packages/contracts/contracts/TestContracts/EBTCTokenTester.sol
@@ -20,7 +20,7 @@ contract EBTCTokenTester is EBTCToken {
         _mint(_account, _amount);
     }
 
-    function unprotectedBurn(address _account, uint _amount) external {
+    function unprotectedBurn(address _account, uint256 _amount) external {
         // No check on caller here
 
         _burn(_account, _amount);
@@ -56,9 +56,9 @@ contract EBTCTokenTester is EBTCToken {
     function getDigest(
         address owner,
         address spender,
-        uint amount,
-        uint nonce,
-        uint deadline
+        uint256 amount,
+        uint256 nonce,
+        uint256 deadline
     ) external view returns (bytes32) {
         return
             keccak256(
