+++ packages/contracts/contracts/TestContracts/LiquityMathTester.sol
@@ -7,17 +7,17 @@ import "../Dependencies/LiquityMath.sol";
 /* Tester contract for math functions in Math.sol library. */
 
 contract LiquityMathTester {
-    function callMax(uint _a, uint _b) external pure returns (uint) {
+    function callMax(uint256 _a, uint256 _b) external pure returns (uint256) {
         return LiquityMath._max(_a, _b);
     }
 
     // Non-view wrapper for gas test
-    function callDecPowTx(uint _base, uint _n) external returns (uint) {
+    function callDecPowTx(uint256 _base, uint256 _n) external returns (uint256) {
         return LiquityMath._decPow(_base, _n);
     }
 
     // External wrapper
-    function callDecPow(uint _base, uint _n) external pure returns (uint) {
+    function callDecPow(uint256 _base, uint256 _n) external pure returns (uint256) {
         return LiquityMath._decPow(_base, _n);
     }
 }
